OPTLGMussoTest/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-exporter-otlp
psycopg2-binary

